This program demonstrates the usage of pthreads, mutexes and qsort(an inbuilt C sorting function).
This programs implements a background thread which sorts a fixed size
array of structures every 10 seconds and displays the array.
The main thread can input elements to write into the fixed size array.
A mutex is used to serialize the array from the main thread and
background thread. An array of structures is used to demonstrate the
usage of a comparator in qsort.