Make the code strictly conforming to ISO C
The code is now strictly conforming to ISO C 90. That also means that it
is forward compatible to newer ISO C standards. Hence, you can compile
it with any conforming implementation.

/* */ comments are accepted in all standards.

The comments and captions are now more accurate.

If `main()` doesn’t accept `argc` and `argv`, it should be `(void)`.

The portable format specifier for pointers is `"%p"` (though the output
differs from one implementation to another). The corresponding argument
must be `(void *)`.