Implementation of Priority Queue using Max Heap Using C Programming Language
I have implemented a priority queue program, which takes the program ID and priority value as inputs, and stores them in a Max-Heap, for efficient querying, in the program. I have implemented this program in the C Programming Language.
In my program, the greater the priority value of the task, the higher is the priority of the task.