Convert an array to reduced form.c
Given an array with n distinct elements, convert the given array to a reduced form where all elements are in the range from 0 to n-1. The order of elements is the same, i.e., 0 is placed in the place of smallest element, 1 is placed for second smallest element, … n-1 is placed for largest element.